Revert "Testing a translate plugin and changing favicon (#2)"
This reverts commit d96b9bacf94e3f7dcc99b65de68ce9ede424b740.
