--- messages/en.json
@@ -1,5 +1,0 @@
-{
-    "index":{
-        "title": "Cool"
-    }
-}
