--- messages/pt.json
@@ -1,5 +1,0 @@
-{
-    "index":{
-        "title": "Daora"
-    }
-}
