+++ next.config.ts
@@ -1,11 +1,7 @@
-import createNextIntlPlugin from 'next-intl/plugin';
 import type { NextConfig } from "next";
 
-const withNextIntl = createNextIntlPlugin();
- 
-/** @type {import('next').NextConfig} */
 const nextConfig: NextConfig = {
   /* config options here */
 };
- 
-export default withNextIntl(nextConfig);
+
+export default nextConfig;
