+++ package.json
@@ -10,7 +10,6 @@
   },
   "dependencies": {
     "next": "^15.0.3",
-    "next-intl": "^3.25.1",
     "react": "19.0.0-rc-66855b96-20241106",
     "react-dom": "19.0.0-rc-66855b96-20241106"
   },
