--- src/i18n/routing.ts
@@ -1,15 +1,0 @@
-import {defineRouting} from 'next-intl/routing';
-import {createNavigation} from 'next-intl/navigation';
- 
-export const routing = defineRouting({
-  // A list of all locales that are supported
-  locales: ['en', 'pt'],
- 
-  // Used when no locale matches
-  defaultLocale: 'pt'
-});
- 
-// Lightweight wrappers around Next.js' navigation APIs
-// that will consider the routing configuration
-export const {Link, redirect, usePathname, useRouter, getPathname} =
-  createNavigation(routing);
